Factor out db fixture
