+++ features/src/test/scala/h4sm/featurerequests/infrastructure/endpoint/RequestEndpointProperties.scala
@@ -1,19 +1,22 @@
 package h4sm.featurerequests
 package infrastructure.endpoint
 
-import cats.effect.IO
+import cats.effect.{ContextShift, IO}
 import h4sm.auth.infrastructure.endpoint.UserRequest
-import db.sql.testTransactor._
 import domain.requests._
 import arbitraries._
+import h4sm.dbtesting.DbFixtureSuite
 import h4sm.featurerequests.db.domain.VotedFeature
 import org.scalatest.prop.PropertyChecks
 
 object RequestEndpointProperties extends PropertyChecks with DbFixtureSuite {
   val dbName = "request_endpoints_test_property_spec"
+  implicit def cs : ContextShift[IO] = IO.contextShift(scala.concurrent.ExecutionContext.Implicits.global)
+  def schemaNames = Seq("ct_auth", "ct_feature_requests")
 
   test("request properties") { p =>
-    import p.reqs._
+    val reqs = new TestRequests[IO](p.transactor)
+    import reqs._
     import authTestEndpoints._
 
     forAll { (feat: FeatureRequest, u: UserRequest) =>
